refactor new user flow
